requirements.txt
asyncpg==0.21.0
click==7.1.2
databases==0.4.1
dotenv-settings-handler==0.0.3
fastapi==0.63.0
gunicorn==20.0.4
h11==0.11.0
psycopg2==2.8.6
pydantic==1.7.3
python-dotenv==0.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.2
